+++ tang-commons/src/main/kotlin/com/tang/commons/utils/queue/config/QueueConfig.kt
@@ -1,0 +1,15 @@
+package com.tang.commons.utils.queue.config
+
+/**
+ * 队列配置
+ *
+ * @author Tang
+ */
+object QueueConfig {
+
+    /**
+     * 槽位数量
+     */
+    const val QUEUE_SIZE: Int = 3600
+
+}